refactor: move applyCrows to its own file
